refactor: moved price selection and wine bottle into one file PriceSelection and removed a space in PairingSelcetion folder
